Use Eugene's new DiskSortedVarIntList

+++ source/Spinach/Helpers/TrigramFileInfo.cs
@@ -10,9 +10,9 @@ public struct TrigramFileInfo : IComparable<TrigramFileInfo>
   {
   }
 
-  public TrigramFileInfo(long fileId, long position)
+  public TrigramFileInfo(ulong fileId, long position)
   {
-    FileId = fileId;
+    FileId = (long)fileId;
     Position = position;
   }
 
@@ -21,6 +21,7 @@ public struct TrigramFileInfo : IComparable<TrigramFileInfo>
   // /////////////////////////////////////////////////////////////////////////////////////////////
 
   public long FileId { get; init; }
+
   public long Position { get; init; }
 
   // /////////////////////////////////////////////////////////////////////////////////////////////
